refactor(types): separate script to generate types
